+++ src/main/java/sreejith/customer/CustomerDao.java
@@ -7,13 +7,13 @@ public interface CustomerDao {
 
     List<Customer> selectAllCustomers();
 
-    Optional<Customer> findCustomerByID(Integer id);
+    Optional<Customer> findCustomerByID(Long id);
 
     void insertCustomer(Customer customer);
 
     Boolean isExistingUser(String email);
 
-    void deleteCustomer(Integer id);
+    void deleteCustomer(Long id);
 
     void updateCustomer(Customer update);
